+++ lib/autokey/gtkui/shared.py
